Use emplace (and emplace_hint) instead of insert
emplace is said to avoid unnecesary copy and move operations. So,
hope that it would avoid them and use them. Though it might not have
given us much of an advantage as we're using just an int.

Also, use emplace_hint to give a hint when we're just filling the set.