count_triplets: try a more optimised solution
For very large inputs sorting the whole array is very costly. So,
let's alternatively try sorting small arrays as the cost is a little
less and might possibly help in improving the performance.

Observation
===========
This pretty much didn't help as only the easy test cases run within
the required time. So, the cost of sorting the elements seems to have
mutiplied. So, this didn't offer quite an improvement.

We would have to try finding some other optimization.