record_linkage: refactor two-loop logic into a single loop via a matching-key map
This refactoring changes the dual loop logic to a single loop
one while maintaining the unique ID in a map whose keys are a
set of matching keys that correspond to that ID.

While at it, refactor the logic into separate classes for better
readability. The map based solution is also better in the aspect
that it represents the solution in a more neat manner hopefully
allowing for easier possibility for further enhancements.